refactor: extract agent machine recovery
